feat: add reusable cocktail ingredient row component
